test(metrics): Merge tests together
